+++ src/lib/utils/taxonomyFilter.ts
@@ -1,8 +1,8 @@
 import { slugify } from "@/lib/utils/textConverter";
 
-const taxonomyFilter = (posts: any[], name: string, key: any) =>
-  posts.filter((post) =>
-    post.data[name].map((name: string) => slugify(name)).includes(key)
+const taxonomyFilter = (communities: any[], name: string, key: any) =>
+  communities.filter((community) =>
+    community.data[name].map((name: string) => slugify(name)).includes(key)
   );
 
 export default taxonomyFilter;