+++ CHANGELOG.md
@@ -7,9 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.0] - 2023-05-20
+
+### Changed
+
+- Rename 'Posts' for 'Communities'
+- Rename 'Authors' for 'Creators'
+- Move components and files to improve order
+
 ## [0.1.0] - 2023-05-17
 
 - initial commit
 
-[unreleased]: https://github.com/ManuelGil/simplus/compare/v0.1.0...HEAD
+[unreleased]: https://github.com/ManuelGil/simplus/compare/v1.0.0...HEAD
+[1.0.0]: https://github.com/ManuelGil/simplus/compare/v0.1.0...1.0.0
 [0.1.0]: https://github.com/ManuelGil/simplus/releases/tag/v0.1.0